+++ src/pymonzo/__init__.py
@@ -14,7 +14,7 @@ from pymonzo.client import MonzoAPI  # noqa
 
 __title__ = "pymonzo"
 __description__ = "Modern Python API client for Monzo public API."
-__version__ = "2.0.0.dev0"
+__version__ = "2.0.0"
 __url__ = "https://github.com/pawelad/pymonzo"
 __author__ = "Paweł Adamczak"
 __email__ = "pawel.ad@gmail.com"