+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## Unreleased
 
+### Fixed
+- Remove Markdown links from PyPI package description.
+
 ## [v1.0.0](https://github.com/pawelad/pymonzo/releases/tag/v1.0.0) - 2024-02-04
 ### Changed
 - Project refresh.