+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog], and this project adheres to
   passed explicitly to `MonzoAPI()` and the settings file couldn't be loaded.
 
 ### Changed
+- Update `MonzoTransactionMerchant` schema with new fields returned by the API.
 - Simplify `MonzoAPI` initialization.
   This (unfortunately) needed an API change because the current attributes (in 
   hindsight) didn't really make sense.